Setting network to regtest directly,
in line with kirill's commits to chromanode. The reason I made regtest a
sub setting of testnet goes back to ngccc, that some components needed
the "testnet" status to be true, for handling bitcoin addresses (since
those components have no knowledge of what a "regtest" formatted address
would be).
However in chromanode all address conversion is handled by bitcore and
it seems the network configuration object for regtest contains enough
information for it to handle regtest addresses, so in chromanode then
there does not seem to be a reason to have a testnet setting being true
while regtest is true.